extract LinkControllerState
