+++ paymentsheet/src/main/java/com/stripe/android/link/LinkControllerViewModel.kt
@@ -26,22 +26,6 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-internal data class LinkControllerState(
-    val linkConfigurationResult: Result<LinkConfiguration?>? = null,
-    val linkGate: LinkGate? = null,
-    val presentedForEmail: String? = null,
-    val linkAccountUpdate: LinkAccountUpdate = LinkAccountUpdate.None,
-    val selectedPaymentMethod: LinkPaymentMethod? = null,
-    val presentPaymentMethodsResult: LinkController.PresentPaymentMethodsResult? = null,
-    val lookupConsumerResult: LinkController.LookupConsumerResult? = null,
-) {
-    val paymentMethodPreview: LinkController.PaymentMethodPreview?
-        get() = (presentPaymentMethodsResult as? LinkController.PresentPaymentMethodsResult.Selected)
-            ?.preview
-
-    val linkConfiguration: LinkConfiguration? = linkConfigurationResult?.getOrNull()
-}
-
 internal class LinkControllerViewModel @Inject constructor(
     application: Application,
     private val paymentElementLoader: PaymentElementLoader,
